adjust code dir
