+++ app/agent/reasoner/dual_model_reasoner.py
@@ -6,13 +6,13 @@ from app.agent.reasoner.model_service import ModelService
 from app.agent.reasoner.model_service_factory import ModelServiceFactory
 from app.agent.reasoner.reasoner import Reasoner
 from app.agent.reasoner.task import Task
-from app.commom.prompt.model_service import TASK_DESCRIPTOR_PROMPT_TEMPLATE
-from app.commom.prompt.reasoner import (
+from app.common.prompt.model_service import TASK_DESCRIPTOR_PROMPT_TEMPLATE
+from app.common.prompt.reasoner import (
     ACTOR_PROMPT_TEMPLATE,
     QUANTUM_THINKER_PROPMT_TEMPLATE,
 )
-from app.commom.system_env import SystemEnv
-from app.commom.type import MessageSourceType
+from app.common.system_env import SystemEnv
+from app.common.type import MessageSourceType
 from app.memory.message import ModelMessage
 from app.memory.reasoner_memory import BuiltinReasonerMemory, ReasonerMemory
 from app.toolkit.tool.tool import Tool
