+++ app/config/sqilte_config.py
@@ -1,6 +1,4 @@
-import os
-
-class Config:
+class SqliteConfig:
     SQLALCHEMY_DATABASE_URI = 'sqlite:///chat2graph.db'
     SQLALCHEMY_TRACK_MODIFICATIONS = False
     DEBUG = True
