+++ app/knowledge/resource/doc.py
