+++ app/memory/message.py
@@ -3,7 +3,7 @@ import time
 from typing import Any, Dict, List, Optional
 from uuid import uuid4
 
-from app.commom.type import MessageSourceType
+from app.common.type import MessageSourceType
 from app.toolkit.tool.tool import FunctionCallResult
