+++ app/plugin/dbgpt/dbgpt_llm_client.py
@@ -16,9 +16,9 @@ from dbgpt.model.proxy.base import LLMClient  # type: ignore
 from dbgpt.model.proxy.llms.chatgpt import OpenAILLMClient  # type: ignore
 
 from app.agent.reasoner.model_service import ModelService
-from app.commom.prompt.model_service import FUNC_CALLING_PROMPT
-from app.commom.system_env import SystemEnv
-from app.commom.type import MessageSourceType
+from app.common.prompt.model_service import FUNC_CALLING_PROMPT
+from app.common.system_env import SystemEnv
+from app.common.type import MessageSourceType
 from app.memory.message import ModelMessage
 from app.toolkit.tool.tool import FunctionCallResult, Tool
