+++ app/service/message_service.py
@@ -1,4 +1,4 @@
-from app.models.sqlite_models import Message, db
+from app.model.sqlite_model import Message, db
 from app.toolkit.api_tool import ParameterException, ServiceException
 from datetime import datetime, timezone
 from openai import OpenAI
