+++ app/toolkit/tool/tool_resource.py
@@ -3,7 +3,7 @@ from typing import Optional
 from uuid import uuid4
 
 from app.agent.reasoner.model_service_factory import ModelServiceFactory
-from app.commom.system_env import SystemEnv
+from app.common.system_env import SystemEnv
 from app.memory.message import ModelMessage
 from app.toolkit.tool.tool import Tool
