+++ example/run_function_calling.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 from uuid import uuid4
 
 from app.agent.reasoner.model_service import ModelService
-from app.commom.type import MessageSourceType
+from app.common.type import MessageSourceType
 from app.memory.message import ModelMessage
 from app.toolkit.tool.tool import Tool
