+++ example/run_model_service.py
@@ -3,8 +3,8 @@ import time
 from typing import List
 
 from app.agent.reasoner.model_service_factory import ModelServiceFactory
-from app.commom.system_env import SystemEnv
-from app.commom.type import MessageSourceType, PlatformType
+from app.common.system_env import SystemEnv
+from app.common.type import MessageSourceType, PlatformType
 from app.memory.message import ModelMessage
