+++ example/run_workflow.py
@@ -7,7 +7,7 @@ from app.agent.reasoner.reasoner import Reasoner
 from app.agent.reasoner.task import Task
 from app.agent.workflow.operator.operator import Operator
 from app.agent.workflow.operator.operator_config import OperatorConfig
-from app.commom.type import WorkflowStatus
+from app.common.type import WorkflowStatus
 from app.memory.message import WorkflowMessage
 from app.memory.reasoner_memory import BuiltinReasonerMemory, ReasonerMemory
 from app.plugin.dbgpt.dbgpt_workflow import DbgptWorkflow
