+++ test/ut/test_dual_model_reasoner.py
@@ -8,7 +8,7 @@ from app.agent.job import Job
 from app.agent.reasoner.dual_model_reasoner import DualModelReasoner
 from app.agent.reasoner.task import Task
 from app.agent.workflow.operator.operator_config import OperatorConfig
-from app.commom.type import MessageSourceType
+from app.common.type import MessageSourceType
 from app.memory.message import ModelMessage
 from app.toolkit.tool.tool import Tool
